feat(footer): refactor footer links structure and update imports
