+++ src/lib/index.ts
@@ -1,2 +1,1 @@
-export * from './footer'
 export * from './utils'
