+++ src/shared/components/footer/footer-links/index.ts
